Viewer groups initial commit

+++ cred/views.py
@@ -192,7 +192,7 @@ def tags(request):
 def detail(request, cred_id):
     cred = get_object_or_404(Cred, pk=cred_id)
 
-    # Check user has perms
+    # Check user has perms as owner or viewer
     if not cred.is_accessible_by(request.user):
         raise Http404
 
@@ -205,10 +205,17 @@ def detail(request, cred_id):
         credlogs = None
         morelink = None
 
+    #User is not in the password owner group, show a read-only UI
+    if not cred.group in request.user.groups.all():
+        readonly = True
+    else:
+        readonly = False
+
     return render(request, 'cred_detail.html', {
         'cred': cred,
         'credlogs': credlogs,
-        'morelink': morelink
+        'morelink': morelink,
+        'readonly': readonly
     })
 
 
@@ -268,7 +275,8 @@ def edit(request, cred_id):
     if request.method == 'POST':
         form = CredForm(request.user, request.POST, request.FILES, instance=cred)
 
-        if form.is_valid():
+        # Password change possible only for owner group
+        if form.is_valid() and cred.group in usergroups:
             # Assume metedata change
             chgtype = CredAudit.CREDMETACHANGE
 
@@ -317,8 +325,8 @@ def delete(request, cred_id):
     except CredAudit.DoesNotExist:
         lastchange = _("Unknown (Logs deleted)")
 
-    # Check user has perms
-    if not cred.is_accessible_by(request.user):
+    # Check user has perms (user must be member of the password owner group)
+    if not cred.is_accessible_by(request.user) or not cred.group in request.user.groups.all():
         raise Http404
     if request.method == 'POST':
         CredAudit(audittype=CredAudit.CREDDELETE, cred=cred, user=request.user).save()
@@ -374,8 +382,8 @@ def tagdelete(request, tag_id):
 @login_required
 def addtoqueue(request, cred_id):
     cred = get_object_or_404(Cred, pk=cred_id)
-    # Check user has perms
-    if not cred.is_accessible_by(request.user):
+    # Check user has perms (user must be member of the password owner group)
+    if not cred.is_accessible_by(request.user) or not cred.group in request.user.groups.all():
         raise Http404
     CredChangeQ.objects.add_to_changeq(cred)
     CredAudit(audittype=CredAudit.CREDSCHEDCHANGE, cred=cred, user=request.user).save()
@@ -386,7 +394,7 @@ def addtoqueue(request, cred_id):
 def bulkdelete(request):
     todel = Cred.objects.filter(id__in=request.POST.getlist('credcheck'))
     for c in todel:
-        if c.is_accessible_by(request.user) and c.latest is None:
+        if c.is_accessible_by(request.user) and c.group in request.user.groups.all() and c.latest is None:
             CredAudit(audittype=CredAudit.CREDDELETE, cred=c, user=request.user).save()
             c.delete()
 
@@ -398,7 +406,7 @@ def bulkdelete(request):
 def bulkundelete(request):
     toundel = Cred.objects.filter(id__in=request.POST.getlist('credcheck'))
     for c in toundel:
-        if c.is_accessible_by(request.user):
+        if c.is_accessible_by(request.user) and c.group in request.user.groups.all():
             CredAudit(audittype=CredAudit.CREDADD, cred=c, user=request.user).save()
             c.is_deleted = False
             c.save()
@@ -411,7 +419,7 @@ def bulkundelete(request):
 def bulkaddtoqueue(request):
     tochange = Cred.objects.filter(id__in=request.POST.getlist('credcheck'))
     for c in tochange:
-        if c.is_accessible_by(request.user) and c.latest is None:
+        if c.is_accessible_by(request.user) and c.group in request.user.groups.all() and c.latest is None:
             CredAudit(audittype=CredAudit.CREDSCHEDCHANGE, cred=c, user=request.user).save()
             CredChangeQ.objects.add_to_changeq(c)
 
@@ -424,7 +432,7 @@ def bulktagcred(request):
     tochange = Cred.objects.filter(id__in=request.POST.getlist('credcheck'))
     tag = get_object_or_404(Tag, pk=request.POST.get('tag'))
     for c in tochange:
-        if c.is_accessible_by(request.user) and c.latest is None:
+        if c.is_accessible_by(request.user) and c.group in request.user.groups.all() and c.latest is None:
             CredAudit(audittype=CredAudit.CREDMETACHANGE, cred=c, user=request.user).save()
             c.tags.add(tag)
